+++ PXCTestKit/CommandLineInterface.swift
@@ -26,11 +26,12 @@ import FBSimulatorControl
                        Option<FileURL>("output", FileURL(url: URL(fileURLWithPath: "output")), description: "Output path where the test reports and log files should be stored."),
                        Option<Locale>("locale", Locale(identifier: "en"), description: "Locale to set for the Simulator."),
                        Option<Preferences>("preferences", Preferences(), description: "Path to some preferences.json to be applied with the Simulator."),
+                       Option<Reporter>("reporter", .rspec, description: "Console reporter type. Supported values: rspec, json"),
                        Option<Only>("only", Only(), description: "Comma separated list of tests that should be executed only. Example: TestClass1/testFoo,TestClass2/testBar"),
                        VaradicOption<Destination>("destination", [], description: "A comma-separated set of key=value pairs describing the destination to use, just like xcodebuild -destination."),
                        Option<Double>("timeout", 3600.0, description: "Timeout in seconds for the test execution to finish."),
                        Flag("no-color", description: "")
-            ) { (testRun, deviceSet, output, locale, preferences, only, destination, timeout, noColor) in
+            ) { (testRun, deviceSet, output, locale, preferences, reporter, only, destination, timeout, noColor) in
                 ANSI.disabled = noColor
                 let consoleOutput = ConsoleOutput()
 
@@ -41,6 +42,7 @@ import FBSimulatorControl
                     locale: locale,
                     environment: ProcessInfo.processInfo.environment,
                     preferences: preferences.dictionary,
+                    reporterType: reporter.type,
                     testsToRun: only.testsToRun,
                     simulators: destination.map({ $0.simulatorConfiguration }),
                     timeout: timeout,